requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
plotly
python-dotenv
groq